+++ CHANGELOG.md
@@ -50,6 +50,7 @@ Line wrap the file at 100 chars.                                              Th
 - Randomize bridge selection with a bias in favor of close bridges.
 - Make login field keep previous value when submitting an incorrect account number in desktop app.
 - Decrease the time it takes to connect to WireGuard relays by sending an ICMP packet immediately.
+- Pause API interactions when the daemon has not been used for 3 days.
 
 ### Fixed
 - Fix the sometimes incorrect time added text after adding time to the account.